created Helper file for player actions
