+++ server/sockets/betting/checkSocket.js
@@ -1,130 +1,12 @@
 const Game = require("../../models/games");
-
-function playersHaveActed(game, currentSeatId, currentAction) {
-  console.log(
-    "[playersHaveActed] Called with currentSeatId:",
-    currentSeatId,
-    "currentAction:",
-    currentAction
-  );
-
-  if (currentAction === "raise") {
-    const result = game.seats.every((seat, index) => {
-      const seatId = seat._id ? seat._id.toString() : "undefined";
-      const condition =
-        !seat.player ||
-        seatId === currentSeatId ||
-        seat.player.action === "fold" ||
-        seat.player.action === "all-in" ||
-        (seat.player.bet >= game.highestBet && seat.player.checkBetFold);
-      console.log(
-        `[playersHaveActed] Seat ${index} (ID: ${seatId}) condition:`,
-        condition,
-        "Player:",
-        seat.player
-      );
-      return condition;
-    });
-    console.log("[playersHaveActed] Returning result:", result);
-    return result;
-  } else {
-    const result = game.seats.every((seat, index) => {
-      const seatId = seat._id ? seat._id.toString() : "undefined";
-      const condition =
-        !seat.player || seatId === currentSeatId || seat.player.checkBetFold;
-      console.log(
-        `[playersHaveActed] Seat ${index} (ID: ${seatId}) condition:`,
-        condition,
-        "Player:",
-        seat.player
-      );
-      return condition;
-    });
-    console.log("[playersHaveActed] Returning result:", result);
-    return result;
-  }
-}
-
-function resetCheckBetFold(game) {
-  game.seats.forEach((seat) => {
-    if (seat.player) {
-      seat.player.checkBetFold = false;
-    }
-  });
-}
-
-function playersWithCards(game) {
-  return game.seats.filter(
-    (seat) => seat.player && seat.player.handCards.length
-  ).length;
-}
-const findNextPosition = (startPosition, seats) => {
-  const seatCount = seats.length;
-  console.log(
-    "[findNextPosition] startPosition:",
-    startPosition,
-    "seatCount:",
-    seatCount
-  );
-
-  let nextPosition = (startPosition + 1) % seatCount;
-  let iterations = 0;
-
-  while (!seats[nextPosition].player) {
-    console.log(
-      "[findNextPosition] Iteration",
-      iterations,
-      "at position",
-      nextPosition,
-      "seat:",
-      seats[nextPosition]
-    );
-    nextPosition = (nextPosition + 1) % seatCount;
-    iterations++;
-    if (iterations > seatCount) {
-      console.error(
-        "[findNextPosition] Could not find a valid seat after",
-        iterations,
-        "iterations."
-      );
-      break;
-    }
-  }
-  console.log("[findNextPosition] Returning nextPosition:", nextPosition);
-  return nextPosition;
-};
-
-function proceedToNextStage(game) {
-  if (game.stage !== "showdown") {
-    if (playersWithCards(game) > 2) {
-      game.stage = "showdown";
-    } else {
-      switch (game.stage) {
-        case "preflop":
-          game.stage = "flop";
-          break;
-        case "flop":
-          game.stage = "turn";
-          break;
-        case "turn":
-          game.stage = "river";
-          break;
-        case "river":
-          game.stage = "showdown";
-          break;
-      }
-    }
-    if (game.stage !== "showdown") {
-      resetCheckBetFold(game);
-    }
-  }
-
-  game.highestBet = 0;
-}
+const {
+  playersHaveActed,
+  findNextPosition,
+  proceedToNextStage,
+} = require("../../utils/actionHelpers");
 
 const checkSocket = (io, socket) => {
   socket.on("check", async (data) => {
-    console.log("[checkSocket] Received check event with data:", data);
     const { gameId, seatId, action } = data;
     try {
       const game = await Game.findById(gameId);
@@ -175,8 +57,6 @@ const checkSocket = (io, socket) => {
         "username"
       );
 
-      // Emit one consolidated event to update all clients.
-      console.log("[checkSocket] Emitting gameUpdated with game:", game);
       io.emit("gameUpdated", updatedGame);
     } catch (error) {
       console.error("[checkSocket] Error handling check:", error);
